Refactor - split off geometry helper functions to separate module
